+++ greenwave/src/app/Providers.jsx
@@ -3,15 +3,15 @@
 import { SessionProvider } from "next-auth/react";
 import { CartProvider } from "./components/cart/cartContext";
 import { UserGlobal } from "./components/users/globalUsers";
-import {SocketProvider} from './websocket'
+import { SocketProvider } from "./websocket";
 export function Providers({ children }) {
   return (
     <SessionProvider>
-      <SocketProvider>
-      <CartProvider>
-        <UserGlobal>{children}</UserGlobal>
-      </CartProvider>
-      </SocketProvider>
+        <SocketProvider>
+          <CartProvider>
+            <UserGlobal>{children}</UserGlobal>
+          </CartProvider>
+        </SocketProvider>
     </SessionProvider>
   );
 }
