--- greenwave/src/app/websocket.js
@@ -1,24 +1,0 @@
-import { useEffect, useState, createContext, useContext } from 'react';
-import io from 'socket.io-client';
-
-export const SocketContext = createContext();
-
-export function SocketProvider({ children }) {
-  const [socket, setSocket] = useState(null);
-
-  useEffect(() => {
-    const newSocket = io.connect('http://localhost:3000');
-    setSocket(newSocket);
-    return () => newSocket.close();
-  }, []);
-
-  return (
-    <SocketContext.Provider value={socket}>
-      {children}
-    </SocketContext.Provider>
-  );
-}
-
-export function useSocket() {
-  return useContext(SocketContext);
-}
